+++ src/editor/editor_client.cpp
@@ -182,7 +182,7 @@ namespace Lux
 
 	void EditorClient::loadUniverse(const char* path)
 	{
-		m_impl->sendMessage(ClientMessageType::LOAD, path, strlen(path)+1);
+		m_impl->sendMessage(ClientMessageType::LOAD, path, (int32_t)strlen(path)+1);
 	}
 
 	void EditorClient::setEntityPosition(int32_t entity, const Vec3& position)
@@ -195,7 +195,7 @@ namespace Lux
 
 	void EditorClient::saveUniverse(const char* path)
 	{
-		m_impl->sendMessage(ClientMessageType::SAVE, path, strlen(path)+1);
+		m_impl->sendMessage(ClientMessageType::SAVE, path, (int32_t)strlen(path)+1);
 	}
 
 	void EditorClient::navigate(float forward, float right, int32_t fast)
